Added the level selector
and the normal functioning is done the locking and unlocking of level if clicked onb level 1 the level 1 scene ( as in build settings) will be loaded.
the functions of the levels are done.
the map making is left and will be done by tonight i think and then the beautification will be left.
if the theme is selected then be sure to notify me